requirements.txt
oauthlib==3.1.0
todoist-python==8.1.2
requests-oauthlib==1.3.0
Flask==1.1.2
tinydb==4.2.0
todoist_service==1.2